Remove drawer component, make parent navigationComponent, remove logic from AppComponent
