+++ src/app/routing/template/app.component.html
@@ -1,17 +1,1 @@
-<pxb-drawer-layout [variant]="getVariant()" (backdropClick)="closeDrawer()">
-    <app-drawer pxb-drawer></app-drawer>
-    <div pxb-content class="drawer-content">
-        <mat-toolbar class="toolbar">
-            <button
-                mat-icon-button
-                *ngIf="variant === 'temporary'"
-                (click)="openDrawer()"
-                style="margin-right: 16px; margin-left: -16px"
-            >
-                <mat-icon>menu</mat-icon>
-            </button>
-            <h2>{{ toolbarTitle }}</h2>
-        </mat-toolbar>
-        <router-outlet></router-outlet>
-    </div>
-</pxb-drawer-layout>
+<app-navigation></app-navigation>
