+++ src/app/routing/template/pages/navigation/home/home.component.scss
