+++ src/app/routing/template/pages/navigation/navigation.component.scss
@@ -1,5 +1,3 @@
-@import '~@pxblue/colors/palette.scss';
-
 .toolbar {
     position: sticky;
     position: -webkit-sticky;
