+++ src/app/routing/template/pages/navigation/page-one/page-one.component.spec.ts
@@ -1,7 +1,7 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { PageOneComponent } from './page-one.component';
-import { AppModule } from '../../app.module';
+import { AppModule } from '../../../app.module';
 
 describe('PageOneComponent', () => {
     let component: PageOneComponent;
