+++ src/app/routing/template/pages/navigation/page-two/page-two.component.scss
